move namespace rnn to a directory
